+++ packages/sdk-ts/src/local/PublicKey.spec.ts
@@ -4,7 +4,7 @@ const privateKey =
   'f9db9bf330e23cb7839039e944adef6e9df447b90b503d5b4464c90bea9022f3'
 
 describe('PublicKey', () => {
-  it('the correct PublicKey is derived from a private key', () => {
+  it('returns the correct PublicKey is derived from a private key', () => {
     const publicKey = PublicKey.fromHex(privateKey)
 
     expect(publicKey.toBase64()).toEqual(
@@ -12,7 +12,15 @@ describe('PublicKey', () => {
     )
   })
 
-  it('the correct Ethereum address is derived from a PublicKey', () => {
+  it('returns the correct PublicKey is derived from a private key starting with 0x', () => {
+    const publicKey = PublicKey.fromHex('0x' + privateKey)
+
+    expect(publicKey.toBase64()).toEqual(
+      'A13cTVZCuTg+Lwh7LuiLcgf2KG68nzEOnfFAbszCwxgT',
+    )
+  })
+
+  it('returns the correct Ethereum address is derived from a PublicKey', () => {
     const publicKey = PublicKey.fromHex(privateKey)
 
     expect(publicKey.toAddress().toHex()).toEqual(
@@ -20,7 +28,7 @@ describe('PublicKey', () => {
     )
   })
 
-  it('the correct Injective address is derived from a PublicKey', () => {
+  it('return the correct Injective address is derived from a PublicKey', () => {
     const publicKey = PublicKey.fromHex(privateKey)
 
     expect(publicKey.toAddress().toBech32()).toEqual(
@@ -28,7 +36,7 @@ describe('PublicKey', () => {
     )
   })
 
-  it('the correct PublicKey is derived from a public key', () => {
+  it('returns the correct PublicKey is derived from a public key', () => {
     const publicKey = PublicKey.fromBase64(
       'A13cTVZCuTg+Lwh7LuiLcgf2KG68nzEOnfFAbszCwxgT',
     )